replace input service
